Move fetch and save features out to Sync class
