+++ src/models/User.ts
@@ -1,4 +1,3 @@
-import axios, { AxiosResponse } from 'axios';
 import { Eventing } from './Eventing';
 
 interface UserProps {
@@ -18,22 +17,5 @@ export class User {
 
 	set(update: UserProps): void {
 		Object.assign(this.data, update);
-	}
-
-	fetch(): void {
-		axios
-			.get(`http://localhost:3000/users/${this.get('id')}`)
-			.then((response: AxiosResponse): void => {
-				this.set(response.data);
-			});
-	}
-
-	save(): void {
-		const id = this.get('id');
-		if (id) {
-			axios.put(`http://localhost:3000/users/${id}`, this.data);
-		} else {
-			axios.post('http://localhost:3000/users', this.data);
-		}
 	}
 }
